requirements.txt
pandas
requests
sqlalchemy
google-cloud-bigquery
pymysql
mysql-connector-python